refactor: extract HasChildSchema trait from ContainerComponent
