+++ src/Core/Components/ContainerComponent.php
@@ -2,35 +2,11 @@
 
 namespace Lattice\Lattice\Core\Components;
 
-use Lattice\Lattice\Attributes\SerializationHook;
-use Lattice\Lattice\Core\Concerns\FiltersRenderableComponents;
+use Lattice\Lattice\Core\Components\Concerns\HasChildSchema;
 
 abstract class ContainerComponent extends Component
 {
-    use FiltersRenderableComponents;
-
-    /**
-     * @var array<int, Component>
-     */
-    protected array $children = [];
-
-    /**
-     * @param  array<int, Component>  $components
-     */
-    public function schema(array $components): static
-    {
-        $this->children = $components;
-
-        return $this;
-    }
-
-    /**
-     * @return array<int, Component>
-     */
-    protected function renderableChildren(): array
-    {
-        return $this->renderableComponents($this->children);
-    }
+    use HasChildSchema;
 
     /**
      * @return array<int, Component>
@@ -48,18 +24,5 @@ abstract class ContainerComponent extends Component
         }
 
         return $result;
-    }
-
-    /**
-     * @param  array<string, mixed>  $data
-     * @return array<string, mixed>
-     */
-    #[SerializationHook(priority: 300)]
-    protected function serialiseSchema(array $data): array
-    {
-        return [
-            ...$data,
-            'schema' => $this->renderableChildren(),
-        ];
     }
 }
